store curr and preloaded tracks in temp

+++ backend/controllers/track.js
@@ -304,8 +304,6 @@ const playTrack = async (req, res) => {
     const patientId = req.query.patientId;
 
     try {
-        // deleteFilesWithPrefix(`${patientId}_`);
-
         const outputFilePath = path.join(
             __dirname,
             "../temp",
@@ -344,14 +342,33 @@ const playTrack = async (req, res) => {
     }
 };
 
-// Delete files in the temp folder with the specified prefix
-const deleteFilesWithPrefix = (prefix) => {
+const cleanTempFolder = (req, res) => {
+    try
+    {
+        const { keepFiles, patientId } = req.body;
+
+        console.log ("CLEANING TEMP FOLDER BUT KEEPING FILES")
+        console.log (keepFiles)
+
+        deleteFilesWithPrefix(`${patientId}_`, keepFiles);
+
+        res.status(200).json({ status: "OK", message: "Temp folder cleaned" });
+    }
+    catch (error)
+    {
+        console.error("Error cleaning temp folder:", error);
+        res.status(500).json({ error: "Error cleaning temp folder" });
+    }
+}
+
+// To clean up the temp folder
+const deleteFilesWithPrefix = (prefix, keepFiles) => {
     const tempFolderPath = path.join(__dirname, "../temp");
     fs.readdir(tempFolderPath, (err, files) => {
         if (err) throw err;
 
         files.forEach((file) => {
-            if (file.startsWith(prefix)) {
+            if (file.startsWith(prefix) && !keepFiles.includes(file)) {
                 fs.unlink(path.join(tempFolderPath, file), (err) => {
                     if (err) throw err;
                 });
@@ -367,5 +384,6 @@ module.exports = {
     scrapeTracks,
     updateTrackRating,
     getTitles,
-    scrapeYtTrack
+    scrapeYtTrack,
+    cleanTempFolder,
 };
